pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deuscode"
version = "0.7.0"
description = "AI-powered multi-agent CLI coding assistant for local LLMs"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.12"
dependencies = [
    "typer",
    "rich",
    "httpx",
    "pyyaml",
    "duckduckgo-search>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
deus = "deuscode.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/deuscode"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
